Using some patterns in getting appium manager
