+++ src/main/java/com/appium/manager/AppiumManagerFactory.java
@@ -1,0 +1,12 @@
+package com.appium.manager;
+
+public class AppiumManagerFactory {
+    public static IAppiumManager getAppiumManager(String host) {
+        if ("127.0.0.1".equals(host)) {
+            return new LocalAppiumManager();
+        } else {
+            return new RemoteAppiumManager();
+        }
+    }
+}
+
