+++ src/main/java/com/appium/manager/IAppiumManager.java
@@ -1,0 +1,12 @@
+package com.appium.manager;
+
+import java.io.IOException;
+
+public interface IAppiumManager {
+
+    void destroyAppiumNode(String host) throws IOException;
+
+    String getRemoteWDHubIP(String host) throws IOException;
+
+    void startAppiumServer(String host) throws Exception;
+}
